pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "docker-compose-sort"
version = "0.0.1"
dependencies = [
    "argparse",
    "ruamel.yaml"
]
authors = [
    { name = "Kyle Johnston", email = "kyle@carabiner.link" }
]
description = "Reorder lines in Docker Compose files to conform to a specific predetermined order"
readme = "README.md"
license = "MIT"
license-files = ["LICENCSE"]
classifiers = [
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Topic :: System :: Systems Administration"
]

[project.urls]
Homepage = "https://github.com/k4j8/docker-compose-sort"
Issues = "https://github.com/k4j8/docker-compose-sort/issues"

[project.scripts]
docker-compose-sort = "docker_compose_sort:main"
